+++ backend/src/main/java/com/almostThere/domain/map/entity/node/MapNode.java
@@ -4,11 +4,11 @@ import lombok.Getter;
 
 @Getter
 public class MapNode {
-    int map_id;
+    long map_id;
     double longitude;
     double latitude;
 
-    public MapNode(int map_id, double longitude, double latitude) {
+    public MapNode(long map_id, double longitude, double latitude) {
         this.map_id = map_id;
         this.longitude = longitude;
         this.latitude = latitude;
